feat(PocketIC): management canister types module (#2140)
This PR introduces a PocketIC library module containing management
canister types to avoid a dependency on the Rust CDK. This is useful to
allow canister developers bump their Rust CDK version independently of
their PocketIC library version. The names of the management canister
types follow the names in the management canister candid interface.