fix: Restore idle timeout functionality for `XNetClientImpl` (#2179)
The old version of hyper's `Client` builder (in `0.14.18`) simply
required a `pool_idle_timeout()` call to set a timeout after which an
idle connection would be dropped from the pool. The new version (in
`hyper-util` `0.1.8`) helpfully preserved the existing API, but broke
the semantics by requiring an additional `pool_timer()` call (to set a
timer) in order for the idle timeout to have any effect whatsoever.

This fix simply adds said `pool_timer()` call.