test(IDX): run //rs/tests/consensus:dual_workload_test in colocated mode to reduce flakiness (#2189)
`//rs/tests/consensus:dual_workload_test` is currently our most flaky
test on PRs. What all flakes have in common is that the testnet is
deployed in `se1` while the driver is running in `zh1` causing cross
Atlantic traffic. What we then see is that some calls the
`http://[$ip]:8080/api/v3/canister/$id/call` time out.

We can potentially fix this by running the test in colocated mode such
that the driver and testnet are located on the same host or DC.