chore: Add graceful shutdown to bin test server (#2198)
The http bin server takes a long time to shutdown at the end of a
system-test.

At the end of a system-test the driver issues a DELETE /group/$testnet
to Farm, Farm will then gracefully shutdown all VMs of the testnet
(using virsh shutdown).

The OS of the UVM running httpbin will then likely send a SIGINT to all
its services including docker to give it time to gracefully shut down.
However, the httpbin doesn't react to SIGINT, so the OS has to wait the
default 90s before it kills the process.