feat(RES-153): deal with Apalache state explosion (#2188)
#1982 added a test with a large number of neurons (>100). While the test
itself (`test_rate_limiting_neuron_creation`) doesn't invoke the TLA
traces check, the current TLA instrumentation implementation causes the
traces of the instrumented methods to be picked up by the other tests in
the same file (it wasn't immediately clear how to improve this). When
the trace of this test is then checked against the TLA model, this
creates large TLA states, which in turn causes Apalache to consume a lot
of time and memory. This PR adds a size gauge to TLA states, and skips
checks of governance transitions involving large states.

I'll work on eliminating the cross-talk between the tests that I
mentioned above, but it seems like a good idea anyway to have this
limit. Maybe we can do something smarter in the future (e.g., maybe the
large strings representing ledger account identifiers are problematic
for Apalache, so maybe we could shorten those), but given that the link
between the code and the model is best-effort, it's acceptable for the
moment IMO.

---------

Co-authored-by: Lara Schmid <73884086+LaraAS@users.noreply.github.com>