feat(Consensus): Enable the `hashes-in-blocks` feature
With the flag set to `true`, we will strip all ingress messages from
blocks, before sending them to peers. On a receiver side, we will
reconstruct the blocks by looking up the referenced ingress messages in
the ingress pool or, if they are not there, by fetching missing ingress
messages from peers who are advertising the blocks.