chore: use WASM chunk store in spec compliance tests (#2251)
This PR uses the WASM chunk store to deploy the universal canister in
spec compliance tests so that the universal canister WASM blob is not
sent raw every time we deploy the universal canister. This optimization
is supposed to save ingress message bandwidth and thereby optimize the
spec compliance test runtime.