chore(PocketIC): use sequence numbers as state labels (#2157)
This PR replaces a state label derived from the certified state by a
sequence number. The reason for this change is that parts of a PocketIC
instance state are not certified (e.g., ingress and canister http pools)
and thus they are not reflected in the state label. Because it is not
clear why having the same state label for two different instances with
the same state would be beneficial at the moment, this PR uses sequence
numbers as state labels.

This PR also fixes a bug when deleting an instance: an instance should
not be deleted if it is still busy with a computation.

Finally,
- the size of the test `//packages/pocket-ic:test` is reduced back to
small (this test is fast again);
- outdated comments are removed;
- unnecessary Debug implementations are removed.

---------

Co-authored-by: IDX GitHub Automation <IDX GitHub Automation>