feat(Index-ng): FI-1389: Disallow upgrading ICRC index-ng from u64 to u256 or vice versa (#1987)
Add a check in the post-upgrade of the ICRC index-ng canister that
causes the canister to trap (and the upgrade to fail) in case an upgrade
attempt is made from a version with `u64` tokens to one with `u256`
tokens or vice versa. Currently, a failure would only be caught after
the post-upgrade, rendering the index-ng canister unusable until it is
upgraded with a fix.