fix(IDX): work around spurious rebuilds in rustix (#2287)
This includes a patch that fixes rebuild issues in `rustix`. Without the
patch the `rustix` crate creates a (seemingly non-deterministic)
`.rmeta` file that causes cache misses in our bazel build.

For more info see:
https://github.com/bytecodealliance/rustix/issues/1199

Some dependencies are also updated to match the `cargo` version and
ensure we only depend on one `rustix` version -- more importantly, that
we don't depend on `rustix v0.37.X` for which the patch doesn't work.