fix: remove the idle future (#2285)
the "idle" method was fairly complex, and it had a few bugs:
- if it was called before any transactions happened, it would not return
immediately, instead it would wait for "idle_sec" (because the last_time
defaults to now()).
- If the last transaction time happened more than "idle_sec" ago, it
would panic because the u64 subtraction would underflow.

It was used to notify the main loop that it needs to break. This can now
be a simple check after each iteration.

These are both very niche cases and almost impossible to happen, from
the way the method is used; but leaving it increases the risk of
misusing the method in the future.