fix(https-outcalls): Increase allowed header size limit for HTTP/2 requests (#2292)
After enabling h/2 with outcalls we had a regression where responses to
https outcalls with headers greated than 16KB would fail. In
https://github.com/hyperium/hyper-util/pull/154#issue-2614460610
hyper-util exposes a method to increase the header response limit.

This PR increases the limit to 52KiB, above 48KiB in the spec, to not
block this weeks release in case the spec-compliance tests have any
headers hardcoded that can break the test. We can reduce this limit
closer to 48KiB in the future.