feat(PSEC-2202): Ignore list for vuln description (#2265)
This PR introduces 'ignore lists' for guestos and boundary guestos which
are applied to vulnerability descriptions of failover findings:
1. Two ignore lists are introduced under `dependencies/resources`.
2. Parser added that reads ignore lists line-by-line, skipping empty
lines and lines that start with `//` which is reserved for comments and
returns the lines in a set.
3. Failover datastore adapted to apply ignore lists, i.e.,
vulnerabilities for findings are not reported if their description
**contains** a line from the ignore list of the corresponding project.


As halloween bonus 🎃 this PR contains code to handle the scary sporadic
error `Expecting value: line 1 column 1 (char 0)` which arises due to
the fact that trivy sometimes fails to fetch their DB from their GH repo
due to rate limiting (see comment in the code).