fix(IDX): bump to typenum v1.17.0 in Bazel (#2296)
The cargo build and the bazel fuzzing builds both use typenum v1.17.0.

The v1.16.0 version used in the main build includes `carg:rustc-env`
stdout lines that include the build directory. Stdout is recorded by
rules_rust and make the `typenum` build artifacts depend on the actual
build directory.

This was fixed in
https://github.com/paholg/typenum/commit/bbac3bf45621b3f35ddd1f312be9a92193470738#diff-aab2aa4b42e7b86d709397b51cfee282c98274effe7e541ed42603b96fe38136L7