feat(sns): Add the upgrade_journal to the get_upgrade_journal SNS-Gov endpoint (#2169)
The idea is for the upgrade journal to have all the information
necessary to understand when and why the SNS got to its current version.
This PR adds the journal to the API (but doesn't yet populate the
journal, generally).

A follow-up PR will use the new endpoint in tests, and populate it in
more places. As a proof of concept, I've made refreshing the upgrade
steps add an entry to the journal, although the purpose of this PR is
just to implement the API.