chore(consensus): Replace generic pool metrics in certification pool (#2312)
This PR replaces the generic artifact pool metrics in the consensus pool
with one that contains height-based info.

I removed the `received_artifact_bytes` for now, because the
implementation is incorrect. We should be relying on `CountBytes` or
some other approach, instead of `size_of_val`. (I also think this kind
of tracking should be defined *once*, for every pool. But that requires
some larger chore work.)