feat: [MR-634] Introduce msg_limit for application subnets in payload builder (#1798)
The `slice.header.begin` of a stream slice we induct is used for garbage
collecting signals. Therefore the number of signals after inducting a
stream slice is `slice.end - slice.header.begin` (note:
`slice.header.begin` <= `slice.begin`). Or in other words, if we want to
cap the number of signals in the reverse stream at `MAX_STREAM_MESSAGES`
(10k), we cannot induct messages beyond `slice.header.begin +
MAX_STREAM_MESSAGES`.

This PR changes the logic of `Payload::take_prefix` such that the above
mentioned limit is respected.

Note that this can not lead to a lock scenario, because a case where we
can't include anymore messages can always be resolved by moving the
`begin` in the incoming `slice`, i.e. by garbage collecting on the
sending subnet.

---------

Co-authored-by: Alin Sinpalean <58422065+alin-at-dfinity@users.noreply.github.com>
Co-authored-by: oggy-dfin <89794951+oggy-dfin@users.noreply.github.com>