fix(recovery): Use metric for highest validated certification for determining node heights (#2332)
During recovery, we care about the highest agreed-upon state. We do this
by fetching node metrics from every node in the subnet. For
certifications, we're currently using the wrong metric.

Instead of using the state manager metric for the last *delivered*
certification, we should fetch the highest certification in the
validated pool.