feat: Update cycles cost for compute (#2308)
This changes the base message execution fee and the per instruction
cycle fee to better capture the strain put on the system given the
situation that was observed with the loaded subnets in the previous
weeks. See also the relevant [forum
post](https://forum.dfinity.org/t/evaluating-compute-pricing-in-response-to-increased-demand-on-the-internet-computer-protocol/36565/8)
for more details. The change for the canister creation fee will follow
in another PR.

---------

Co-authored-by: Mathias Björkqvist <mathias.bjorkqvist@dfinity.org>
Co-authored-by: Arshavir Ter-Gabrielyan <arshavir.ter.gabrielyan@dfinity.org>