refactor(cketh/ckerc20): modular log scraping (#2258)
Refactor the ckETH minter log scraping to be more modular, meaning that
the fetching and the parsing of Ethereum logs is separated.

 For a given log scraping, it is known at compile-time:

1. How to retrieve the Ethereum logs, which is an implementation of
`LogScraping`
2. How to parse the retrieved Ethereum logs, which is an implementation
of `LogParser`.

This provides the following benefits:

1. Easier to add a new log scraping.
2. Dedicated log parser for each log scraping ensuring that events
between log scrapings cannot be mixed.
3. Use exactly the same logic to drive the log scrapings, which includes
the concept of a chunkable block range, needed since most providers
constraints the width of the block range that can be used when
retrieving Ethereum logs.

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>