feat(nns): Added Neuron::voting_power_refreshed_timestamp_seconds field. (#2268)
When creating a new neuron, this field gets set to the same value as
created_timestamp_seconds.

For existing neurons, they will be treated as though the value in this
field is 1 second after midnight 2024-11-05 UTC (until they refresh).

In a later PR, the value in this field will get updated to the current
time whenever the user performs "relevant" operations.

Closes https://dfinity.atlassian.net/browse/NNS1-3404.

[Next PR 👉][next]

[next]: https://github.com/dfinity/ic/pull/2320