fix: fix calculating reference instruction cost in tests (#2334)
This PR corrects the calculation of reference instruction costs in
execution tests.

This code is for testing purposes only and is not part of production. 
The issue went unnoticed earlier because we mainly assert on cycle costs
rather than instruction costs, and on smaller values, this difference
was lost during instructions-to-cycles conversion.