chore(system-tests): adding API BNs to ic-prep and creating a testnet with API BNs (#2290)
This PR adds the API BNs to the system test infrastructure such that one
can spin up an IC with API boundary nodes. To this end, `ic-prep` has
been extended (e.g., domain name in the node structs and the notion of
API boundary nodes). In addition, the test driver has been extended to
allow the addition of API boundary nodes. Finally, a small testnet with
API BNs has been added to serve as an example.

In follow-up MRs, we can move some of the system tests (e.g., the API BN
decentralization test).

---------

Co-authored-by: IDX GitLab Automation <idx@dfinity.org>