chore(consensus): adjust dynamic block maker delay (#2356)
The current value is constant and a bit too high. It should depend on
the size of the subnet, because on a subnet with fewer nodes it should
take less time to broadcast a block than on a subnet with many nodes.
For example, on an app subnet it should take around 3x less time than on
an nns subnet.

The unit delay setting in the subnet record encapsulates quite well what
we need, thus instead of having the constant, we use the unit delay.
This effectively means that when a subnet struggles to produce rank-0
blocks at acceptable (at the moment 66%) rate we double the time we
allot to a rank-0 block maker for a creation and distribution of a
block.