feat: RUN-1039: Convert executed instructions to cycles differently for Wasm64 (#2291)
In this PR the conversion from executed instructions to cycles is
treated differently for Wasm64 compared to Wasm32. This PR puts in place
all the scaffolding and infrastructure that allows Wasm64 to be treated
differently.

This PR inspects the Wasm on the replica side before `install_code` to
determine what kind of charging it should apply.
In follow-up work we will either (1) keep this process as-is but then
pass on the unzipped Wasm to the sandbox; (2) Call into the sandbox to
get this information.

The most important part of the change is related to
`cycles_account_manager/lib.rs`, `install.rs`, `install_code.rs`,
`update.rs`, all the other changes are mostly mechanical changes.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>
Co-authored-by: Dimitris Sarlis <dimitrios.sarlis@dfinity.org>