chore: remove unused clippy allows (#2371)
There are two important changes in the MR.

- Directives like `#[allow(unused)]` or `#[allow(dead_code)]` are fairly
dangerous if misplaces. Because clippy will skip further analysis of
potential problems in the corresponding scopes. One example is futures
that are not poll. This is a correctness bug that people rely on the
compiler to detect this. Having `allow(unused) or allow(dead_code)` may
result in bad changes going to production because clippy was told to
ignore them implicitly.
- I removed a test with `bincode::deserialize_from` since we don't have
any usage `bincode::deserialize_from` in code base