refactor(ckbtc): refactor the ic-btc-kyt crate (#2364)
Refactor the ic-btc-kyt crate to only keep necessary functions in the
library and move the rest to build together with the Wasm binary.

This is a necessary move in order to resolve a bizarre problem with CI
when ic-btc-kyt is made a dependency of ic-ckbtc-minter. More
specifically, our CI job "Bazel Build All Config Check" (which is
essentially running clippy) would fail unless askama template is moved
out of the ic-btc-kyt lib. Refactoring this crate avoids the problem.