feat: Add an `mr_receive_batch_latency_seconds` metric
The metric tracks the latency of the `Receiver::recv()` call that `MessageRoutingImpl` makes to obtain the next batch from Consensus.

On a subnet with a block rate of 2.5 blocks/s whose DSM does zero work and finishes immediately, we should consistently observe values around 400ms. If OTOH the DSM is backlogged and always takes more than 400ms to run, then the observed values should be consistently around zero (because the DSM would be consistently behind Consensus).