test(IDX): move //rs/embedders:stable_grow_test to its own dedicated test and give it more CPUs
The `//rs/embedders:embedders_integration_tests/stable_grow` test often
times out (at 5m) on CI when the runner is scheduled to a loaded
node. The P90 at the time of writing is 4m 53s.

Let's move this test to a dedicated rust_ic_test and reserve a bit
more CPUs to see if that reduces the overall parallelism and allows
this test to run more quickly.