feat(nns): Make distinction between potential and deciding voting power. (#2339)
Add deciding_voting_power method to Neuron.

Renames the voting_power method of Neuron to potential_voting_power.

Added a flag that controls whether deciding_voting_power applies
adjustment. Per usual, adjustment is off in vanilla builds, but on when
feature = "test".

Voting rewards uses this new method. As long as the the new behavior is
not enabled (via the flag), this should result in no behavior change,
and even initially, when the feature is enabled, since no neuron can be
too unrefreshed, even then, there would be no immediate behavior change.

# Background

This is part of the "periodic confirmation" feature that was recently
approved in proposal [132411][prop].

[prop]: https://dashboard.internetcomputer.org/proposal/132411

# References

Closes https://dfinity.atlassian.net/browse/NNS1-3406,
https://dfinity.atlassian.net/browse/NNS1-3407.

[👈 Previous PR][prev] | [Next PR 👉][next]

[prev]: https://github.com/dfinity/ic/pull/2338
[next]: https://github.com/dfinity/ic/pull/2375