chore(CRP-2617): Generalize CUP’s ChainKeyInitialization (#2337)
This PR extends the `ChainKeyInitialization` with a variant for
`InitialNiDkgTranscriptRecord`. This is needed to support subnets with
VetKeys.

Note on compatibility:
We are moving a single field of `ChainKeyInitialization` into a `oneof`,
that has another variant.
Since `oneof`s in Protobuf are simply invariants that checks that only
one of the specified fields is populated, this does not change the wire
representation of this type, if the `initialization` is of variant
`dealings` (See [protobuf language
guide](https://protobuf.dev/programming-guides/proto3/#backward)).

This code then simply adapts the client code (NNS registry and
consensus) to reflect the changes. Since we don't introduce any code
that creates the newly introduced `transcript` variant, this code does
not introduce incompatibilities, so no migration code is necessary.

The changes for the registry that make use of the new `transcript`
variant are planned in a later PR, but a number of pieces have to be in
place first, such as the management endpoints and the consensus code.

---------

Co-authored-by: Arshavir Ter-Gabrielyan <arshavir.ter.gabrielyan@dfinity.org>