feat: [MR-632] Protect in-memory states for CUP creation and validation (#2223)
Previous PR: https://github.com/dfinity/ic/pull/2061

This is the second PR towards removing in-memory states more eagerly at
checkpointed heights.
Currently, the in-memory state at the previous checkpoint height is kept
until the next CUP, roughly the whole checkpointing interval. We plan to
remove it more eagerly and only keep it until the current CUP.

Currently, `remove_inmemory_states_below` purges states based on the
certified height referenced in the finalized tip. As the certified
height could go beyond the summary block height, we need to keep certain
states for several rounds more until the current CUP is created. Thus
Consensus needs to inform the state manager of the extra heights to keep
through a new API .

This PR uses the new API with proper arguments, i.e. filling in all the
heights required for CUP creation and validation, and adds an end-to-end
test unit test. Specifically, we keep the in-memory states of all
Summary blocks with IDKG payload, for which there exists no CUP artifact
yet. These states are needed to create and validate CUP shares on IDKG
subnets (see
[here](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/consensus/src/consensus/catchup_package_maker.rs?L212-227)
and
[here](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/consensus/src/consensus/validator.rs?L1677-1683)).

---------

Co-authored-by: Leo Eichhorn <leo.eichhorn@dfinity.org>
Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>