feat(consensus): CON-675 Increase notary delay caused by certified-finalized gap (#2450)
Recently, we are observing multiple subnets maintaining a (relatively)
constant gap of 3-5 rounds between their latest finalized and certified
heights, which likely leads to higher end-to-end ingress latencies on
these subnets.

With this PR, we attempt to reduce the size of this gap, by increasing
the notary delay more aggressively once the "acceptable gap" between
finalization and certification is exceeded. This is done by including
the heights _within_ the acceptable gap when calculating the additive
linear delay.

Additionally, instead of using the unit delay (1s on app subnets, 3s on
large/system subnets) to calculate the delay, we switch to a new
"backlog delay" which is set to a constant 2 seconds. The unit delay
isn't related to the speed of execution, so having two values for it is
confusing.

Furthermore, we 
- reduce the acceptable gap from 3 to 2,
- fix a test to calculate the expected delay correctly.

Running the `consensus_performance_test` with an added execution delay
of 2s gives the following results:
- Certified-finalized gap: 4\~5 (old implementation) -> 2\~3 (new
implementation)
- Average E2E ingress latency: 10.2s (old implementation) -> 8.8s (new
implementation)

![chart
(1)](https://github.com/user-attachments/assets/882eeed2-9ab4-47be-8b24-c4f9aded5033)