feat(nns): Added total_potential_voting_power to proposals. (#2375)
# Background

This is part of the "periodic confirmation" feature that was recently
approved in proposal [132411][prop].

[prop]: https://dashboard.internetcomputer.org/proposal/132411

# Future Work

This will be used in the calculation of voting rewards. Whereas, before,
the total voting power in ballots (i.e. "deciding" voting power) was
used as the baseline to determine each neuron's share of the reward
purse. With the distinction between potential voting power vs. deciding
voting power (implemented in an earlier PR in this series), if we
continued to do it the old way, the pie would be sliced into a smaller
number of pieces. When this new field is used, the pie will continue to
be sliced into the same number of pieces today. This way, neurons that
stay refreshed will keep getting the same amount of rewards (not more),
while those who do not stay refreshed will get less.

# References

Closes https://dfinity.atlassian.net/browse/NNS1-3411.

[👈 Previous PR][prev] | [Next PR 👉][next]

[prev]: https://github.com/dfinity/ic/pull/2339
[next]: https://github.com/dfinity/ic/pull/2385