feat(cketh/ckerc20): consolidate log scrapings (#2449)
Follow-up on #2324 and #2369 to consolidate log scrapings done by the
ckETH minter.

### Added

Information related to the state of the scraping of the new deposit with
subaccount helper smart contract:

1. Last scraped block number to `MinterInfo`.
2. State of scraping (address and last scraped block number) to the
minter dashboard.
3. Metric for the last scraped block number.

### Changed

Refactor `State` to have a container (`LogScrapings`) to aggregate all
log scrapings. This allows to:

1. Automatically display the state of a new log scraping to the minter
dashboard. Information related to helper smart contracts was moved to
its own section.
2. Automatically add a metric for the last scraped block number.
3. Recall to add an event that is registered during `pre-upgrade` to
save the last scraped block number to stable memory.