chore: add idleness bitcoin adapter integration tests (#2389)
Two tests were added:
1. The adapter should not establish any connections if it starts off as
inactive / idle (which is the case for all subnets not running the
bitcoin feature)
2. The adapter should close all connections once it becomes idle. 
> I approached this by decreasing the idleness threshold to only a few
seconds and sleeping. There are tests that take considerably longer than
the new ones anyway
> `tokio::time::pause` and `tokio::time::advance` don't work because the
setup is multithreaded

One thing to note is that actual connections to peers happen during a
tick / heartbeat; meaning that if the adapter is initialized and then
awaits indefinitely (without going through a tick) it's technically an
OK behaviour, as no downloads / connections would occur. Meaning that
https://github.com/dfinity/ic/pull/2278 was not necessary. However these
are still good tests.