chore(IDX): optimise the setup of the spec_compliance systests (#2482)
This optimises the setup function of the spec_compliance system-tests by
concurrently installing the NNS and starting the Boundary Node. This
shaves off a minute from the setup and brings the total test time under
5 minutes.

Additionally the steps on the critical path of the setup function are
now measured and logged so that we can monitor what takes up the time.