chore(idkg): Replace pseudo-random and pre-signature IDs in `RequestId` with single `CallbackId` (#2054)
The `RequestId` currently contains the `PseudoRandomId` and the
`PreSigId` of the corresponding context. Instead, we can use the
`CallbackId` of the context directly, which would allow us to re-use the
`RequestId` also for vet KD requests, which don't have a
`PseudoRandomId` or `PreSigId`.

Using the `CallbackId` instead is safe, because all three IDs
(`PseudoRandomId`, `PreSigId` and `CallbackId`) uniquely identify a call
context. Looking forward, we should remove the `PseudoRandomId`
entirely. Historically, it was used as an identifier, and as a source of
randomness for signature generation. Today, the randomness comes from a
dedicated `nonce` field instead, which means the `PseudoRandomId` is no
longer needed.

Changing the protobuf definition of `RequestId` and immediately using it
is safe because, unlike the CUP, these artifacts are deleted on every
subnet upgrade. Therefore, they do not have to be backwards or forwards
compatible.