feat(BOUN-1236): implement secret tracking (#2433)
This change adds a `Tracker` implementation for the client of the
canister. This allows one to instantiate a tracker that will follow the
current secret value from the canister, as well as participate in leader
duties for secret generation (i.e if you as a client have been chosen as
a leader, you will be required to generate the salt and encrypt it for
other peers).