feat(BOUN-1236): add logging and throttling to canister methods (#2434)
This is a fairly minimal change that adds throttling and logging to the
canister methods in the client. Note we use
`#[allow(clippy::disallowed_types)]` to not warn on using
`tokio::sync::Mutex` as it's in the disallowed types list in
`clippy.toml`. The reason we _do_ need to use the tokio mutex here is
because it is being used in the throttling code and it needs to hold on
to the lock until the throttling finishes (i.e it's not a quick
lock/unlock, but rather can take a while).