feat: Enable jemalloc profiling
This builds jemalloc with support for profiling enabled. A bit more work is required for something like a `/pprpf/heap` endpoint. And IDX will probably not like whatever it is that actually building jemalloc from scratch does.

In order to check that this works, run (probably with `CARGO_BAZEL_REPIN=true` prepended the first time around):

```
_RJEM_MALLOC_CONF=prof:true,prof_final:true,prof_leak:true,prof_gdump:true,lg_prof_interval:32,prof_prefix:/tmp/jeprof bazel run //rs/replica:replica
```

You should end up with a bunch of `/tmp/jeprof*` files that can be viewed with the `jeprof` tool.

For more context see https://www.magiroux.com/rust-jemalloc-profiling/